fix season type mismatch in pick join

+++ server/routes/trade-intelligence.ts
@@ -58,7 +58,7 @@ async function getAssetsForTrades(
       JOIN league_chain lc_trade ON lc_trade.league_id = ta.league_id
       JOIN league_chain lc_draft ON lc_draft.root_id = lc_trade.root_id AND lc_draft.league_id = dr.league_id
       LEFT JOIN draft_traded_picks dtp ON dtp.league_id = dr.league_id
-        AND dtp.season = split_part(ta.asset_key, '_', 1)::int
+        AND dtp.season = split_part(ta.asset_key, '_', 1)
         AND dtp.round = split_part(ta.asset_key, '_', 2)::int
         AND dtp.original_owner_id = split_part(ta.asset_key, '_', 3)::int
       WHERE dr.season = split_part(ta.asset_key, '_', 1)
